+++ docs/src/partials/home/companies.js
@@ -1,18 +1,16 @@
 import React from "react";
 import styled from "styled-components";
-
-import LazyImage from "../../partials/lazy-image";
-import createPath from "../../helpers/path-helpers";
+import Image from "next/image";
 
 // Assets
-import VIACOM from "../../../static/logos/logo-viacom.svg";
-import FIVETHIRTYEIGHT from "../../../static/logos/logo-fivethirtyeight.svg";
-import USAFACTS from "../../../static/logos/logo-usafacts.svg";
-import AIRBNB from "../../../static/logos/logo-airbnb.svg";
-import REDFIN from "../../../static/logos/logo-redfin.svg";
-import TUNE from "../../../static/logos/logo-tune.svg";
-import ZILLOW from "../../../static/logos/logo-zillow.svg";
-import BENAROYA from "../../../static/logos/logo-benaroya.png";
+import VIACOM from "../../static/logos/logo-viacom.svg";
+import FIVETHIRTYEIGHT from "../../static/logos/logo-fivethirtyeight.svg";
+import USAFACTS from "../../static/logos/logo-usafacts.svg";
+import AIRBNB from "../../static/logos/logo-airbnb.svg";
+import REDFIN from "../../static/logos/logo-redfin.svg";
+import TUNE from "../../static/logos/logo-tune.svg";
+import ZILLOW from "../../static/logos/logo-zillow.svg";
+import BENAROYA from "../../static/logos/logo-benaroya.png";
 import importedTheme from "../../styles/theme";
 import {
   LinkButton,
@@ -53,9 +51,7 @@ const Company = styled.li`
   max-width: 200px;
 `;
 
-const CompanyLogo = styled(LazyImage)`
-  width: 100%;
-`;
+const CompanyLogo = Image;
 
 const Companies = () => (
   <LandingSectionWrapper bg={importedTheme.color.lightGray}>
@@ -63,34 +59,35 @@ const Companies = () => (
       <SectionHeading>A Few of Our Fans</SectionHeading>
       <CompaniesList>
         <Company>
-          <CompanyLogo src={AIRBNB} alt="Airbnb" />
+          <CompanyLogo src={AIRBNB} alt="Airbnb" title="Airbnb" />
         </Company>
         <Company>
-          <CompanyLogo src={FIVETHIRTYEIGHT} alt="FiveThirtyEight" />
+          <CompanyLogo
+            src={FIVETHIRTYEIGHT}
+            alt="FiveThirtyEight"
+            title="FiveThirtyEight"
+          />
         </Company>
         <Company>
-          <CompanyLogo src={REDFIN} alt="Redfin" />
+          <CompanyLogo src={REDFIN} alt="Redfin" title="Redfin" />
         </Company>
         <Company>
-          <CompanyLogo src={USAFACTS} alt="USAFacts" />
+          <CompanyLogo src={USAFACTS} alt="USAFacts" title="USAFacts" />
         </Company>
         <Company>
-          <CompanyLogo src={VIACOM} alt="Viacom" />
+          <CompanyLogo src={VIACOM} alt="Viacom" title="Viacom" />
         </Company>
         <Company>
-          <CompanyLogo src={BENAROYA} alt="Benaroya" />
+          <CompanyLogo src={BENAROYA} alt="Benaroya" title="Benaroya" />
         </Company>
         <Company>
-          <CompanyLogo src={ZILLOW} alt="Zillow" />
+          <CompanyLogo src={ZILLOW} alt="Zillow" title="Zillow" />
         </Company>
         <Company>
-          <CompanyLogo src={TUNE} alt="Tune" />
+          <CompanyLogo src={TUNE} alt="Tune" title="Tune" />
         </Company>
       </CompaniesList>
-      <LinkButton
-        to={createPath("about#showcase")}
-        color={importedTheme.color.black}
-      >
+      <LinkButton href="/about#showcase" className="mx-auto bg-white">
         SEE SHOWCASE
       </LinkButton>
     </LandingSectionContent>