+++ docs/src/partials/sidebar/styles/index.js
@@ -2,7 +2,6 @@ import SidebarSectionHeading from "./sidebar-section-heading";
 import SidebarSectionList from "./sidebar-section-subheading";
 import SidebarSectionSublist from "./sidebar-section-sublist";
 import SidebarListItem from "./sidebar-list-item";
-import SidebarListItemAnchorLink from "./sidebar-list-item-anchor-link";
 import SidebarListItemLink from "./sidebar-list-item-link";
 
 export {
@@ -10,6 +9,5 @@ export {
   SidebarSectionList,
   SidebarSectionSublist,
   SidebarListItem,
-  SidebarListItemAnchorLink,
   SidebarListItemLink,
 };