+++ packages/victory-native/src/components/victory-cursor-container.tsx
@@ -42,7 +42,7 @@ function nativeCursorMixin<
     };
 
     // overrides all web events with native specific events
-    static defaultEvents = (props: TProps) => {
+    static defaultEvents(props: TProps) {
       return [
         {
           target: "parent",
@@ -65,7 +65,7 @@ function nativeCursorMixin<
           },
         },
       ];
-    };
+    }
   };
 }