+++ packages/victory-scatter/CHANGELOG.md
@@ -1,5 +1,29 @@
 # victory-scatter
 
+## 37.0.1
+
+### Patch Changes
+
+- Fix the signature of class static functions in components ([#2840](https://github.com/FormidableLabs/victory/pull/2840))
+
+## 37.0.0
+
+### Major Changes
+
+- Upgrade babel dependencies and build target to modern browsers ([#2804](https://github.com/FormidableLabs/victory/pull/2804))
+
+## 36.9.2
+
+### Patch Changes
+
+- Replace lodash array utils with native code ([#2810](https://github.com/FormidableLabs/victory/pull/2810))
+
+* Replace lodash values and mapValues with native code ([#2808](https://github.com/FormidableLabs/victory/pull/2808))
+
+- Replace lodash isNil and isNan with native code ([#2800](https://github.com/FormidableLabs/victory/pull/2800))
+
+* Replace lodash isFunction with native code ([#2802](https://github.com/FormidableLabs/victory/pull/2802))
+
 ## 36.9.1
 
 ## 36.9.0