+++ packages/victory-vendor/CHANGELOG.md
@@ -1,5 +1,15 @@
 # victory-vendor
 
+## 37.0.1
+
+## 37.0.0
+
+### Major Changes
+
+- Upgrade babel dependencies and build target to modern browsers ([#2804](https://github.com/FormidableLabs/victory/pull/2804))
+
+## 36.9.2
+
 ## 36.9.1
 
 ## 36.9.0